Fix English in 'debounce()' description
There's an important difference between restrictive ('which') and nonrestrictive relative clauses ('that') in English. 

Since all versions of debounce meet the description given, 'which' is the correct word. 

Additionally this means people who don't know what a debounced function is won't stop reading (currently it implies that some unknown things have a certain behvaior, rather than all unknown things having a certain behavior, ie, right now it's not a definition, with 'which' it would be).  