Fix usage of hash function in `_.memoize`.
The function `_.memoize` has a bug where the first argument to the
function is used as the key when retrieving a result from the cache,
ignoring the hash function.

The fix is always to use the computed `address` as the key to the cache.

Confirmed the regression and fix with the updated unit test.