Loosen clock skew check on debounce
In 4596a7e1, a sanity check was added to debounce to make sure our clock hadn't skewed earlier. If the debounced function is getting called under load, though, the last called ts can be the same as the current ts when the timeout runs. That causes `last` to be 0, and currently causes the debounced function to fire.