Test: Use same QUnit version in CI as on website
Follows-up 351875092c35044 and d317fdda606b40f2d3f16.

Contrary to the commit message, it actually restored the older 2.6 version,
and not updated to the newer QUnit 2.10 version from package.json.

The oldest Node.js LTS is currently Node 12.
The latest QUnit 2.18 supports Node 10+ still.
This repo currently has a Node 8 CI job still.

While at it, update both to 2.10.1, the last version to support Node 8,
and pin it to avoid "npm install" from expanding it, and to avoid
e.g. using a newer method which then fails on the website later.