Add nativeDataView check to isDataView
The is DataView check relies on DataView being a constructor like
it is natively. However, if an application overrides DataView from
the window this ends up throwing an error when underscore is
included. This here will check if DataView is the native implementation
before trying to use the constructor on it. If it is not native
it will fallback on the same heuristic used on older IE versions
for checking if something is a DataView or not.